+++ app/src/main/java/com/hbv2/dlf_plus/networks/websocket/WSChatClient.kt
@@ -12,6 +12,7 @@ import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.schedulers.Schedulers
 import ua.naiksoftware.stomp.Stomp
 import ua.naiksoftware.stomp.StompClient
+import ua.naiksoftware.stomp.dto.LifecycleEvent
 import java.io.IOException
 import java.util.concurrent.Callable
 import java.util.function.Function
@@ -19,25 +20,45 @@ import java.util.function.Function
 class WSChatClient {
     private lateinit var mStompClient: StompClient
     private var compositeDisposable = CompositeDisposable()
-    var messagees = ArrayList<MessageDTO>()
-    init {
+    init { //sleppa þessu, þurfum bara stomp.over, setja það í onCreate.
         if(!::mStompClient.isInitialized){
             mStompClient = Stomp.over(Stomp.ConnectionProvider.OKHTTP, Constants.WS_URL)
-            mStompClient.connect()
         }
     }
-    fun closeConnection(){
+    fun connect(){ //connecta þegar message er opnað
+        resetSubscriptions()
+        mStompClient.connect()
+        val stompConnection = mStompClient.lifecycle().subscribe {
+            when (it.type) {
+                LifecycleEvent.Type.OPENED -> {
+                    Log.d("stompConn", "OPENED")
+                }
+                LifecycleEvent.Type.CLOSED ->{
+                    Log.d("stompConn","CLOSED")
+                }
+                LifecycleEvent.Type.ERROR -> {
+                    Log.d("stompConn","ERROR")
+                }
+                else -> {
+                    Log.d("stompConn","Something went wrong")
+                }
+            }
+        }
+        compositeDisposable.add(stompConnection)
+        mStompClient.connect()
+    }
+    fun closeConnection(){ //loka þegar hætt er í message activity
         if(::mStompClient.isInitialized && mStompClient.isConnected){
             mStompClient.disconnect()
         }
         compositeDisposable.dispose();
     }
-    fun subscribe(threadId: Int, addItem: (msg: MessageDTO) -> Boolean){
+    fun subscribe(threadId: Int){ // senda skilaboð. addItem
         val subscription = mStompClient.topic("/thread/$threadId/get")
             .subscribeOn(Schedulers.io())
             .observeOn(AndroidSchedulers.mainThread())
             .subscribe{
-                addItem(Gson().fromJson(it.payload,MessageDTO::class.java))
+                //addItem(Gson().fromJson(it.payload,MessageDTO::class.java))
             }
         compositeDisposable.add(subscription)
     }
